refactor(checker): shard dispatch.rs this-keyword and JSX handlers into dedicated modules
dispatch.rs was at 1,988 lines — 12 lines from the §19 hard limit of
2000. Any new handler added would immediately violate the rule. Extract
the two largest inline match-arm bodies:

- ThisKeyword (~297 lines) → dispatch_this.rs
- JSX_ELEMENT (~234 lines) → dispatch_jsx.rs

dispatch.rs drops from 1,988 → 1,457 lines. Both new modules follow
the existing dispatch_helpers.rs / dispatch_yield.rs companion pattern:
`pub(super) fn` on the same `ExpressionDispatcher<'a, 'b>` impl, with
delegate calls in the match arms of dispatch.rs.

Zero behavior change. Verified: `cargo check -p tsz-checker` passes.
