+++ crates/tsz-checker/src/dispatch.rs
@@ -87,303 +87,7 @@ impl<'a, 'b> ExpressionDispatcher<'a, 'b> {
             k if k == SyntaxKind::RegularExpressionLiteral as u16 => {
                 self.dispatch_regular_expression_literal(idx)
             }
-            k if k == SyntaxKind::ThisKeyword as u16 => {
-                {
-                    use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                    // TS2465: 'this' cannot be referenced in a computed property name.
-                    // Check this first — it takes priority over other `this` errors.
-                    if self
-                        .checker
-                        .is_this_in_class_member_computed_property_name(idx)
-                    {
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::THIS_CANNOT_BE_REFERENCED_IN_A_COMPUTED_PROPERTY_NAME,
-                            diagnostic_codes::THIS_CANNOT_BE_REFERENCED_IN_A_COMPUTED_PROPERTY_NAME,
-                        );
-                        return TypeId::ANY;
-                    }
-                    // TS2332: `this` inside enum member initializers is always invalid,
-                    // even when the enum is nested in a namespace.
-                    if self.checker.is_this_in_enum_member_initializer(idx) {
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::THIS_CANNOT_BE_REFERENCED_IN_CURRENT_LOCATION,
-                            diagnostic_codes::THIS_CANNOT_BE_REFERENCED_IN_CURRENT_LOCATION,
-                        );
-                        // tsc emits the companion TS2683 when `this` is directly in
-                        // the enum initializer. When `this` is captured through an
-                        // arrow function, TS2683 is NOT emitted (the arrow's `this`
-                        // captures the outer context which is the enum — still invalid
-                        // via TS2332, but not flagged for implicit-any).
-                        if !self.checker.has_enclosing_arrow_before_enum(idx) {
-                            self.checker.error_at_node(
-                                idx,
-                                diagnostic_messages::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                                diagnostic_codes::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                            );
-                        }
-                        return TypeId::ANY;
-                    }
-                    // TS2331: 'this' cannot be referenced in a module or namespace body
-                    // In JS files, `namespace` is invalid syntax (TS8006) so tsc
-                    // doesn't emit TS2331/TS2683 for `this` in namespace bodies.
-                    if !self.checker.is_js_file() && self.checker.is_this_in_namespace_body(idx) {
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::THIS_CANNOT_BE_REFERENCED_IN_A_MODULE_OR_NAMESPACE_BODY,
-                            diagnostic_codes::THIS_CANNOT_BE_REFERENCED_IN_A_MODULE_OR_NAMESPACE_BODY,
-                        );
-                        // TSC always emits TS2683 as a companion to TS2331 in
-                        // namespace bodies — `this` is inherently untyped here,
-                        // regardless of noImplicitThis.
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                            diagnostic_codes::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                        );
-                        return TypeId::ANY;
-                    }
-                    // TS17009: 'super' must be called before accessing 'this'
-                    if self
-                        .checker
-                        .is_this_before_super_in_derived_constructor(idx)
-                    {
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::SUPER_MUST_BE_CALLED_BEFORE_ACCESSING_THIS_IN_THE_CONSTRUCTOR_OF_A_DERIVED_CLASS,
-                            diagnostic_codes::SUPER_MUST_BE_CALLED_BEFORE_ACCESSING_THIS_IN_THE_CONSTRUCTOR_OF_A_DERIVED_CLASS,
-                        );
-                    }
-                    // TS2816: Cannot use 'this' in a static property initializer of a decorated class
-                    if self.checker.ctx.compiler_options.experimental_decorators
-                        && let Some(ref class_info) = self.checker.ctx.enclosing_class
-                        && class_info.in_static_property_initializer
-                        && !self.checker.is_this_in_nested_function_inside_class(idx)
-                        && let Some(class_node) = self.checker.ctx.arena.get(class_info.class_idx)
-                        && let Some(class_data) = self.checker.ctx.arena.get_class(class_node)
-                        && let Some(ref modifiers) = class_data.modifiers
-                    {
-                        let has_class_decorator = modifiers.nodes.iter().any(|&mod_idx| {
-                            self.checker.ctx.arena.get(mod_idx).is_some_and(|n| {
-                                n.kind == tsz_parser::parser::syntax_kind_ext::DECORATOR
-                            })
-                        });
-                        if has_class_decorator {
-                            self.checker.error_at_node(
-                                            idx,
-                                            diagnostic_messages::CANNOT_USE_THIS_IN_A_STATIC_PROPERTY_INITIALIZER_OF_A_DECORATED_CLASS,
-                                            diagnostic_codes::CANNOT_USE_THIS_IN_A_STATIC_PROPERTY_INITIALIZER_OF_A_DECORATED_CLASS,
-                            );
-                        }
-                    }
-                }
-                let has_intermediate_function = self
-                    .checker
-                    .is_this_in_nested_function_without_own_this_binding(idx);
-                let contextual_owner = self.checker.this_has_contextual_owner(idx);
-                let contextual_owner_is_class = contextual_owner
-                    .and_then(|owner_idx| self.checker.ctx.arena.get(owner_idx))
-                    .is_some_and(|owner_node| {
-                        owner_node.kind == syntax_kind_ext::CLASS_DECLARATION
-                            || owner_node.kind == syntax_kind_ext::CLASS_EXPRESSION
-                    });
-                if !has_intermediate_function
-                    && let Some(member_idx) = self.checker.find_enclosing_function(idx)
-                    && let Some(member_node) = self.checker.ctx.arena.get(member_idx)
-                    && (member_node.kind == syntax_kind_ext::METHOD_DECLARATION
-                        || member_node.kind == syntax_kind_ext::GET_ACCESSOR
-                        || member_node.kind == syntax_kind_ext::SET_ACCESSOR
-                        || member_node.kind == syntax_kind_ext::CONSTRUCTOR)
-                    && let Some(class_member_this_type) =
-                        self.checker.class_member_this_type(member_idx)
-                {
-                    return self
-                        .checker
-                        .apply_flow_narrowing(idx, class_member_this_type);
-                }
-                if self.checker.is_js_file()
-                    && let Some(func_idx) = self.checker.find_enclosing_non_arrow_function(idx)
-                    && let Some(jsdoc) = self.checker.get_jsdoc_for_function(func_idx)
-                    && let Some(this_expr) =
-                        CheckerState::extract_jsdoc_tag_type_expression(&jsdoc, "this")
-                    && let Some(this_type) = self.checker.resolve_jsdoc_reference(this_expr)
-                {
-                    return self.checker.apply_flow_narrowing(idx, this_type);
-                }
-                if let Some(this_type) = self.checker.current_this_type() {
-                    let transient_this_marker =
-                        crate::query_boundaries::property_access::is_this_type(
-                            self.checker.ctx.types,
-                            this_type,
-                        ) || crate::query_boundaries::common::contains_type_parameters(
-                            self.checker.ctx.types,
-                            this_type,
-                        );
-                    // A nested regular function creates its own `this` binding.
-                    // Ignore any outer contextual/class `this` unless the
-                    // function itself owns that binding.
-                    if !has_intermediate_function
-                        && (!contextual_owner_is_class || !transient_this_marker)
-                    {
-                        return self.checker.apply_flow_narrowing(idx, this_type);
-                    }
-                    // Fall through — either the nested function has its own `this`,
-                    // or the lexical owner is a class member whose `this` binding
-                    // should outrank transient call-context ThisType markers.
-                }
-                if let Some(class_idx) = self
-                    .checker
-                    .ctx
-                    .enclosing_class
-                    .as_ref()
-                    .map(|info| info.class_idx)
-                    .or_else(|| {
-                        (!has_intermediate_function
-                            && !contextual_owner.is_some_and(|owner_idx| {
-                                self.checker.ctx.arena.get(owner_idx).is_some_and(|owner| {
-                                    owner.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION
-                                })
-                            }))
-                        .then(|| self.checker.nearest_enclosing_class(idx))
-                        .flatten()
-                    })
-                {
-                    // Inside a class but no explicit this type on stack -
-                    // return the class instance/constructor type depending on static context.
-                    // BUT: if `this` is inside a nested regular function (not a class member),
-                    // that function creates its own `this` binding, so don't use the class type.
-                    // Walk the AST to determine static context — can't rely on
-                    // in_static_member flag since it's only set during check_class_member.
-                    let is_in_static = self.checker.is_in_static_class_member_context(idx);
-                    if !has_intermediate_function {
-                        if let Some(class_node) = self.checker.ctx.arena.get(class_idx)
-                            && let Some(class_data) = self.checker.ctx.arena.get_class(class_node)
-                        {
-                            let this_type = if is_in_static {
-                                self.checker
-                                    .get_class_constructor_type(class_idx, class_data)
-                            } else {
-                                self.checker.get_class_instance_type(class_idx, class_data)
-                            };
-                            return self.checker.apply_flow_narrowing(idx, this_type);
-                        }
-                        TypeId::ANY
-                    } else {
-                        // Fall through to TS2683 / TS7041 checks below
-                        // Suppress if the nested function has an explicit `this` parameter
-                        // or a contextual `this` type from a parent type annotation
-                        if self.checker.ctx.no_implicit_this()
-                            && !self
-                                .checker
-                                .enclosing_function_has_explicit_this_parameter(idx)
-                            && !self
-                                .checker
-                                .enclosing_function_has_contextual_this_type(idx)
-                        {
-                            use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                            self.checker.error_at_node(
-                                idx,
-                                diagnostic_messages::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                                diagnostic_codes::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                            );
-                        }
-                        TypeId::ANY
-                    }
-                } else if self.checker.this_has_contextual_owner(idx).is_some()
-                    && !self
-                        .checker
-                        .is_this_in_nested_function_without_own_this_binding(idx)
-                {
-                    self.checker.ctx.recover_any(
-                        idx,
-                        crate::recovery::RecoveryReason::ThisUnresolvedClassOrObjectLiteralMember,
-                    )
-                } else if self.checker.ctx.no_implicit_this()
-                    && !self.checker.is_js_file()
-                    && !self.checker.ctx.binder.is_external_module()
-                    && self.checker.is_this_in_global_capturing_arrow(idx)
-                {
-                    // TS7041: `this` in an arrow chain with no enclosing
-                    // function/class/object `this` binder captures globalThis.
-                    // Prefer this over the generic TS2683 path.
-                    //
-                    // Only fires in *script* files. At the top level of an
-                    // external module (has `import`/`export`), `this` is
-                    // `undefined`, not `globalThis`, so a capturing arrow at
-                    // module top-level produces TS2532 on property access,
-                    // not a global-capture warning. Matches tsc.
-                    use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                    self.checker.error_at_node(
-                        idx,
-                        diagnostic_messages::THE_CONTAINING_ARROW_FUNCTION_CAPTURES_THE_GLOBAL_VALUE_OF_THIS,
-                        diagnostic_codes::THE_CONTAINING_ARROW_FUNCTION_CAPTURES_THE_GLOBAL_VALUE_OF_THIS,
-                    );
-                    TypeId::ANY
-                } else if self.checker.ctx.no_implicit_this()
-                    && self
-                        .checker
-                        .find_enclosing_non_arrow_function(idx)
-                        .is_some()
-                    && (!self.checker.is_js_file()
-                        || self
-                            .checker
-                            .is_this_in_nested_function_without_own_this_binding(idx))
-                {
-                    // TS2683: 'this' implicitly has type 'any'
-                    // In JS files, only nested regular functions with a fresh,
-                    // unowned `this` binding reach this path. Constructor/prototype
-                    // patterns and explicit/contextual/JSDoc-owned receivers are
-                    // filtered out before this branch.
-                    // Suppress if the enclosing function has an explicit `this` parameter
-                    // or a contextual `this` type from a parent type annotation
-                    if self
-                        .checker
-                        .enclosing_function_has_explicit_this_parameter(idx)
-                        || self
-                            .checker
-                            .enclosing_function_has_contextual_this_type(idx)
-                    {
-                        TypeId::ANY
-                    } else {
-                        use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                        self.checker.error_at_node(
-                            idx,
-                            diagnostic_messages::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                            diagnostic_codes::THIS_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_TYPE_ANNOTATION,
-                        );
-                        TypeId::ANY
-                    }
-                } else if self.checker.ctx.no_implicit_this()
-                    && !self.checker.is_js_file()
-                    && self
-                        .checker
-                        .find_enclosing_non_arrow_function(idx)
-                        .is_none()
-                {
-                    // `this` at the top level of a script or module with noImplicitThis.
-                    //
-                    // In an external module (has `import`/`export`), top-level `this`
-                    // — including `this` inside a top-level arrow — is `undefined`.
-                    // Property access on `this` then produces TS2532 ("Object is
-                    // possibly 'undefined'.") under strictNullChecks, matching tsc.
-                    //
-                    // In a script, tsc resolves `this` to `typeof globalThis` (an
-                    // object type). We approximate with TypeId::OBJECT since we
-                    // don't have a full globalThis type yet; this ensures that
-                    // operations like `++this` correctly emit TS2356 (arithmetic
-                    // type error) instead of TS2357 (invalid lvalue) — matching
-                    // tsc behavior where the type check fires first and suppresses
-                    // the lvalue check.
-                    if self.checker.ctx.binder.is_external_module() {
-                        TypeId::UNDEFINED
-                    } else {
-                        TypeId::OBJECT
-                    }
-                } else {
-                    TypeId::ANY
-                }
-            }
+            k if k == SyntaxKind::ThisKeyword as u16 => self.dispatch_this_keyword(idx),
             k if k == SyntaxKind::SuperKeyword as u16 => {
                 self.checker.get_type_of_super_keyword(idx)
             }
@@ -1488,241 +1192,7 @@ impl<'a, 'b> ExpressionDispatcher<'a, 'b> {
                 TypeId::VOID
             }
             // JSX Elements (Rule #36: JSX Intrinsic Lookup)
-            k if k == syntax_kind_ext::JSX_ELEMENT => {
-                if let Some(jsx) = self.checker.ctx.arena.get_jsx_element(node) {
-                    // Extract contextual type for children from the component's
-                    // `children` prop BEFORE processing children, so arrow functions
-                    // and other expressions get contextual parameter typing.
-                    let children_ctx_type = if !jsx.children.nodes.is_empty() {
-                        self.checker
-                            .get_jsx_children_contextual_type(jsx.opening_element)
-                    } else {
-                        None
-                    };
-                    let children_request = request
-                        .read()
-                        .normal_origin()
-                        .contextual_opt(children_ctx_type);
-                    // Collect children types for children prop synthesis.
-                    // tsc synthesizes a `children` prop from JSX element body children
-                    // and validates it against the component's `children` prop type.
-                    let mut child_types: Vec<TypeId> = Vec::new();
-                    let mut has_text_child = false;
-                    let mut text_child_indices: Vec<NodeIndex> = Vec::new();
-                    let mut has_spread_child = false;
-                    for &child in &jsx.children.nodes {
-                        if let Some(child_node) = self.checker.ctx.arena.get(child) {
-                            // Skip trivial whitespace JsxText — tsc ignores whitespace-only
-                            // text that contains newlines (formatting indentation). But
-                            // same-line whitespace (e.g., `<A />  <B />`) is preserved.
-                            if child_node.kind == tsz_scanner::SyntaxKind::JsxText as u16
-                                && let Some(text) = self.checker.ctx.arena.get_jsx_text(child_node)
-                            {
-                                let is_all_whitespace =
-                                    text.text.chars().all(|c| c.is_ascii_whitespace());
-                                let has_newline = text.text.contains('\n');
-                                if is_all_whitespace && has_newline {
-                                    continue;
-                                }
-                            }
-                            // Skip empty JSX expressions (e.g., {/* comment */})
-                            // — tsc does not count these as children.
-                            if child_node.kind == syntax_kind_ext::JSX_EXPRESSION
-                                && let Some(expr_data) =
-                                    self.checker.ctx.arena.get_jsx_expression(child_node)
-                                && expr_data.expression == NodeIndex::NONE
-                            {
-                                continue;
-                            }
-                        }
-                        let child_type = if let Some(child_node) = self.checker.ctx.arena.get(child)
-                            && child_node.kind == syntax_kind_ext::JSX_EXPRESSION
-                            && let Some(expr_data) =
-                                self.checker.ctx.arena.get_jsx_expression(child_node)
-                            && expr_data.dot_dot_dot_token
-                        {
-                            has_spread_child = true;
-                            let spread_type = self.checker.get_type_of_node_with_request(
-                                expr_data.expression,
-                                &children_request,
-                            );
-                            self.checker
-                                .normalize_jsx_spread_child_type(child, spread_type)
-                        } else if let Some(child_node) = self.checker.ctx.arena.get(child)
-                            && child_node.kind == syntax_kind_ext::JSX_EXPRESSION
-                            && let Some(expr_data) =
-                                self.checker.ctx.arena.get_jsx_expression(child_node)
-                            && expr_data.expression.is_some()
-                            && self
-                                .checker
-                                .ctx
-                                .arena
-                                .get(expr_data.expression)
-                                .is_some_and(|expr| {
-                                    matches!(
-                                        expr.kind,
-                                        syntax_kind_ext::ARROW_FUNCTION
-                                            | syntax_kind_ext::FUNCTION_EXPRESSION
-                                    )
-                                })
-                        {
-                            let has_function_context =
-                                children_request.contextual_type.is_some_and(|ctx_type| {
-                                    let ctx_type =
-                                        self.checker.resolve_type_for_property_access(ctx_type);
-                                    crate::query_boundaries::common::function_shape_for_type(
-                                        self.checker.ctx.types,
-                                        ctx_type,
-                                    )
-                                    .is_some()
-                                        || crate::query_boundaries::common::call_signatures_for_type(
-                                            self.checker.ctx.types,
-                                            ctx_type,
-                                        )
-                                        .is_some_and(|sigs| !sigs.is_empty())
-                                });
-                            if has_function_context {
-                                self.checker
-                                    .ctx
-                                    .implicit_any_contextual_closures
-                                    .insert(expr_data.expression);
-                                self.checker
-                                    .ctx
-                                    .implicit_any_checked_closures
-                                    .insert(expr_data.expression);
-                                self.checker.invalidate_function_like_for_contextual_retry(
-                                    expr_data.expression,
-                                );
-                                self.checker.get_type_of_node_with_request(
-                                    expr_data.expression,
-                                    &children_request,
-                                )
-                            } else {
-                                // Still process the arrow/function expression to emit
-                                // diagnostics (e.g., TS7006 for unannotated parameters).
-                                // Without a callable contextual type, the result is ANY,
-                                // but we must evaluate the function to trigger TS7006.
-                                self.checker.get_type_of_node(child);
-                                TypeId::ANY
-                            }
-                        } else {
-                            self.checker
-                                .get_type_of_node_with_request(child, &children_request)
-                        };
-                        if let Some(child_node) = self.checker.ctx.arena.get(child)
-                            && child_node.kind == tsz_scanner::SyntaxKind::JsxText as u16
-                        {
-                            has_text_child = true;
-                            text_child_indices.push(child);
-                        }
-                        child_types.push(child_type);
-                    }
-                    // Synthesize the children type:
-                    // - 0 children → None (no children prop synthesized)
-                    // - 1 child → the child's type directly
-                    // - 2+ children without spread → fixed tuple of child types
-                    // - 2+ children with spread → array of union of child types
-                    let children_ctx = if !child_types.is_empty() {
-                        let child_count = child_types.len();
-                        let synthesized_type = if child_types.len() == 1 && !has_spread_child {
-                            child_types[0]
-                        } else if !has_spread_child {
-                            let elements = child_types
-                                .iter()
-                                .copied()
-                                .map(|type_id| tsz_solver::TupleElement {
-                                    type_id,
-                                    name: None,
-                                    optional: false,
-                                    rest: false,
-                                })
-                                .collect();
-                            self.checker.ctx.types.factory().tuple(elements)
-                        } else {
-                            // Multiple children: synthesize as an array type.
-                            // tsc uses the union of all child types as the element type.
-                            let element_type =
-                                self.checker.ctx.types.factory().union(child_types.clone());
-                            self.checker.ctx.types.factory().array(element_type)
-                        };
-                        let normalized_child_count = if has_spread_child {
-                            child_count.max(2)
-                        } else {
-                            child_count
-                        };
-                        Some(crate::checkers_domain::JsxChildrenContext {
-                            child_count: normalized_child_count,
-                            has_text_child,
-                            contextual_type: children_ctx_type,
-                            synthesized_type,
-                            text_child_indices,
-                        })
-                    } else {
-                        None
-                    };
-                    // Check closing element for TS7026 (tsc emits for both opening and closing tags)
-                    self.checker
-                        .check_jsx_closing_element_for_implicit_any(jsx.closing_element);
-                    // tsc also re-checks the closing tag's name identifier, so an
-                    // unresolved component name emits TS2304 at both the opening
-                    // and closing positions. Route the closing tag_name through
-                    // type resolution so the standard identifier-lookup path
-                    // produces the diagnostic.
-                    if let Some(closing_node) = self.checker.ctx.arena.get(jsx.closing_element)
-                        && let Some(closing_data) =
-                            self.checker.ctx.arena.get_jsx_closing(closing_node)
-                    {
-                        let tag_name_idx = closing_data.tag_name;
-                        if let Some(tag_name_node) = self.checker.ctx.arena.get(tag_name_idx)
-                            && tag_name_node.kind == tsz_scanner::SyntaxKind::Identifier as u16
-                            && let Some(ident) =
-                                self.checker.ctx.arena.get_identifier(tag_name_node)
-                            && ident
-                                .escaped_text
-                                .chars()
-                                .next()
-                                .is_some_and(|c| c.is_ascii_uppercase())
-                        {
-                            let _ = self
-                                .checker
-                                .get_type_of_node_with_request(tag_name_idx, request);
-                        }
-                        if let Some(tag_name_node) = self.checker.ctx.arena.get(tag_name_idx)
-                            && tag_name_node.kind == syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION
-                        {
-                            let mut root_idx = tag_name_idx;
-                            while let Some(root_node) = self.checker.ctx.arena.get(root_idx)
-                                && root_node.kind == syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION
-                                && let Some(access) =
-                                    self.checker.ctx.arena.get_access_expr(root_node)
-                            {
-                                root_idx = access.expression;
-                            }
-                            if let Some(root_node) = self.checker.ctx.arena.get(root_idx)
-                                && root_node.kind == tsz_scanner::SyntaxKind::Identifier as u16
-                                && let Some(ident) =
-                                    self.checker.ctx.arena.get_identifier(root_node)
-                                && ident
-                                    .escaped_text
-                                    .chars()
-                                    .next()
-                                    .is_some_and(|c| c.is_ascii_uppercase())
-                            {
-                                let _ = self
-                                    .checker
-                                    .get_type_of_node_with_request(root_idx, request);
-                            }
-                        }
-                    }
-                    self.checker.get_type_of_jsx_opening_element_with_children(
-                        jsx.opening_element,
-                        request,
-                        children_ctx,
-                    )
-                } else {
-                    TypeId::ERROR
-                }
-            }
+            k if k == syntax_kind_ext::JSX_ELEMENT => self.dispatch_jsx_element(idx, request),
             k if k == syntax_kind_ext::JSX_SELF_CLOSING_ELEMENT => self
                 .checker
                 .get_type_of_jsx_opening_element_with_children(idx, request, None),
